pyproject.toml
[project]
name = "certik-vasp-scope"
version = "1.0.0"
description = "IN 701 VASP scoping API + static SPA"
requires-python = ">=3.12"
# Obrigatório para a Vercel instalar pacotes ao usar pyproject.toml
# (https://vercel.com/docs/functions/runtimes/python#dependencies)
dependencies = [
    "fastapi>=0.109.0,<1",
    "starlette>=0.27.0,<1",
    "pydantic>=2.0.0,<3",
    "typing-extensions>=4.8.0,<5",
    "PyYAML>=6.0.1,<7",
    "pandas>=2.0.0,<3",
    "google-generativeai>=0.8.0,<1",
    "tabulate>=0.9.0,<0.10",
]

# https://vercel.com/docs/frameworks/backend/fastapi
[project.scripts]
app = "main:app"

# Manter [project.dependencies] alinhado com o núcleo em requirements.txt (CI valida com scripts/verify_pyproject_runtime_deps.py).
